Move `all`
